+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/homeserver/HomeServerCapabilities.kt
@@ -48,6 +48,15 @@ data class HomeServerCapabilities(
         UNKNOWN
     }
 
+    /**
+     * Check if a feature is supported by the homeserver.
+     * @return
+     *  UNKNOWN if the server does not implement room caps
+     *  UNSUPPORTED if this feature is not supported
+     *  SUPPORTED if this feature is supported by a stable version
+     *  SUPPORTED_UNSTABLE if this feature is supported by an unstable version
+     *  (unstable version should only be used for dev/experimental purpose)
+     */
     fun isFeatureSupported(feature: String): RoomCapabilitySupport {
         if (roomVersions?.capabilities == null) return RoomCapabilitySupport.UNKNOWN
         val info = roomVersions.capabilities[feature] ?: return RoomCapabilitySupport.UNSUPPORTED
@@ -74,6 +83,12 @@ data class HomeServerCapabilities(
         return info.preferred == byRoomVersion || info.support.contains(byRoomVersion)
     }
 
+    /**
+     * Use this method to know if you should force a version when creating
+     * a room that requires this feature.
+     * You can also use #isFeatureSupported prior to this call to check if the
+     * feature is supported and report some feedback to user.
+     */
     fun versionOverrideForFeature(feature: String) : String? {
         val cap = roomVersions?.capabilities?.get(feature)
         return cap?.preferred ?: cap?.support?.lastOrNull()
