+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/Room.kt
@@ -101,8 +101,4 @@ interface Room :
      * Use this room as a Space, if the type is correct.
      */
     fun asSpace(): Space?
-
-    suspend fun setJoinRulePublic()
-    suspend fun setJoinRuleInviteOnly()
-    suspend fun setJoinRuleRestricted(allowList: List<String>)
 }
