+++ vector/src/main/java/im/vector/app/features/roomprofile/settings/joinrule/advanced/RoomJoinRuleChooseRestrictedViewModel.kt
@@ -188,6 +188,7 @@ class RoomJoinRuleChooseRestrictedViewModel @AssistedInject constructor(
 
     fun handleSubmit() = withState { state ->
         setState { copy(updatingStatus = Loading()) }
+
         viewModelScope.launch {
             try {
                 when (state.currentRoomJoinRules) {
